Extracted Bullet into standalone component
- Added stories for Bullet component
